awsome pawsome update

+++ src/main/java/com/example/wmsdsktp/Responses/ProdutoResponse.java
@@ -7,7 +7,7 @@ public record ProdutoResponse(
         String unidade,
         Double preco,
         Double aproxVolume,
-        Integer utilizadorCriadorId
+        Long utilizadorCriadorId
 ) {
 
 }
